+++ src/commands/enrolSupernode.ts
@@ -18,10 +18,10 @@ import { Command } from '@oclif/command';
 import { BootstrapService, BootstrapUtils } from '../service';
 import { AnnounceService } from '../service/AnnounceService';
 
-export default class Supernode extends Command {
-    static description = `It registers the nodes in the supernode rewards program by announcing the enrol transaction to the registration address.`;
+export default class EnrolSupernode extends Command {
+    static description = `It enrols the nodes in the supernode rewards program by announcing the enrol transaction to the registration address.`;
 
-    static examples = [`$ symbol-bootstrap supernode`];
+    static examples = [`$ symbol-bootstrap enrolSupernode`];
 
     static flags = {
         help: BootstrapUtils.helpFlag,
@@ -30,8 +30,8 @@ export default class Supernode extends Command {
     };
 
     public async run(): Promise<void> {
-        const { flags } = this.parse(Supernode);
+        const { flags } = this.parse(EnrolSupernode);
         BootstrapUtils.showBanner();
-        return new BootstrapService(this.config.root).supernode(flags);
+        return new BootstrapService(this.config.root).enrolSupernode(flags);
     }
 }